+++ app/src/main/java/ru/pavelcoder/modulbankdemo/model/bank/Transaction.kt
@@ -3,6 +3,6 @@ package ru.pavelcoder.modulbankdemo.model.bank
 data class Transaction(
     val sourceCurrency: Currency,
     val destinationCurrency: Currency,
-    val sourceAmount: Long,
-    val destinationAmount: Long
+    val sourceAmount: Double,
+    val destinationAmount: Double
 )
